user action creators merged into general action creators for tables
