+++ client/src/redux/actions/GET/a_getLoggedBooks.js
@@ -13,7 +13,7 @@ import { LIVE_BE_base_URL } from '../../../utils'
 
 
     // Action Creator
-    export const a_getLoggedBooks = () => {
+    export const a_getLoggedBooks = (userID) => {
     // console.log('INSIDE: a_getLoggedBooks action creator')
     // -- //
         // Send First Action
@@ -33,12 +33,18 @@ import { LIVE_BE_base_URL } from '../../../utils'
             // Make Axios Request
             axios
                 .get(used_URL)
-                .then(completedBooks => {
-                // console.log(completedBooks)
+                .then(allLogs => {
+                // console.log(allLogs)
                 // -- // 
+                    const userResults = allLogs.data.filter(item => item.userID === userID)
+                    console.log(userResults)
+
                     dispatch({
                         type: GET_ALLLOGGEDBOOKS_SUCCESS,
-                        payload: completedBooks.data
+                        payload: {
+                            allUsers: allLogs.data,
+                            singleUser: userResults
+                        }
                     })
                 })
                 .catch(err => {
