+++ client/src/redux/reducers/r_reviews.js
@@ -3,10 +3,6 @@ import {
 } from '../actions/GET/a_getReviews.js'
 
 import {
-    GET_USERREVIEWS_START, GET_USERREVIEWS_SUCCESS, GET_USERREVIEWS_FAILURE
-} from '../actions/GET/a_getUserReviews.js'
-
-import {
     ADD_REVIEW_START, ADD_REVIEW_SUCCESS, ADD_REVIEW_FAILURE
 } from '../actions/POST/a_addReview.js'
 
@@ -43,9 +39,10 @@ export const r_reviews = (state=initialState, action) => {
             return {
                 ...state,
                 
-                DB_reviews: [...action.payload],
+                DB_reviews: [...action.payload.allUsers],
+                USER_reviews: [...action.payload.singleUser],
+                
                 is_fetching: false,
-
                 error: ''
             }
 
@@ -56,28 +53,28 @@ export const r_reviews = (state=initialState, action) => {
                 is_fetching: false,
             }
 
-        // - 2 - // get SINGLE USERS reviews
-        case GET_USERREVIEWS_START:
-            return {
-                ...state,
-                is_fetchingUserData: true,
-                error: '',
-            }
-        case GET_USERREVIEWS_SUCCESS:
-            return {
-                ...state,
+        // // - 2 - // get SINGLE USERS reviews
+        // case GET_USERREVIEWS_START:
+        //     return {
+        //         ...state,
+        //         is_fetchingUserData: true,
+        //         error: '',
+        //     }
+        // case GET_USERREVIEWS_SUCCESS:
+        //     return {
+        //         ...state,
 
-                USER_reviews: [...action.payload],
-                is_fetchingUserData: false,
+        //         USER_reviews: [...action.payload],
+        //         is_fetchingUserData: false,
 
-                error: '',
-            }
-        case GET_USERREVIEWS_FAILURE:
-            return {
-                ...state,
-                error: action.payload,
-                is_fetchingUserData: false,
-            }
+        //         error: '',
+        //     }
+        // case GET_USERREVIEWS_FAILURE:
+        //     return {
+        //         ...state,
+        //         error: action.payload,
+        //         is_fetchingUserData: false,
+        //     }
 
         // - 3 - // Add Review
         case ADD_REVIEW_START:
@@ -92,8 +89,8 @@ export const r_reviews = (state=initialState, action) => {
 
                 DB_reviews: [...action.payload.allUsers],
                 USER_reviews: [...action.payload.singleUser],
-                is_adding: false,
                 
+                is_adding: false,
                 error: ''
             }
         case ADD_REVIEW_FAILURE:
