+++ client/src/views/homepage.jsx
@@ -16,11 +16,13 @@ import ListSelector from '../components/homepage/explore_NYT_booksAPI/ListSelect
 import VertTabPannel from '../components/homepage/explore_NYT_booksAPI/VertTabPannel.js'
 
 // ACTION CREATORS
-import { a_getSpecificList } from '../redux/actions/GET/a_getSpecificList.js'
 import { a_getBooks } from '../redux/actions/GET/a_getBooks.js'
 import { a_getReviews } from '../redux/actions/GET/a_getReviews.js'
 import { a_getLoggedBooks } from '../redux/actions/GET/a_getLoggedBooks.js'
 import { a_getRanks } from  '../redux/actions/GET/a_getRanks.js'
+
+// FUNCTIONS
+import decode from '../utils/decode_JWT.js'
               
 // === === === === === === === === === === === === //
 // === === === === === === === === === === === === //
@@ -40,8 +42,9 @@ const useStyles = makeStyles(theme => ({
 // -B- COMPONENT
 function HomePage(props) {
 const { 
+    token,
     allLists, listName, searchDate,
-    a_getBooks, a_getReviews, a_getLoggedBooks, a_getRanks     // Action Creators
+    a_getBooks, a_getReviews, a_getLoggedBooks, a_getRanks,
 } = props
     // Styles
     const classes = useStyles({})
@@ -52,10 +55,15 @@ const {
         We now want to load initial data from your DB into the redux store
     */
     useEffect(() => {
+        // Get User ID
+            const userID = decode(token).user_ID
+            console.log(userID)
+
+        // Get Data
         a_getBooks()
-        a_getReviews()
-        a_getLoggedBooks()
-        a_getRanks()
+        a_getLoggedBooks(userID)
+        a_getReviews(userID)
+        a_getRanks(userID)
     }, [])
 
     // Return
@@ -88,6 +96,8 @@ const mstp = state => {
         searchDate: state.r_specificList.searchDate,
 
         current_listData: state.r_specificList.listData,
+
+        token: state.r_auth.token,
     }
 }
         
@@ -96,8 +106,6 @@ export default connect(
     mstp,
     {
         a_getBooks,
-        a_getReviews,
-        a_getLoggedBooks,
-        a_getRanks,
+        a_getLoggedBooks, a_getReviews, a_getRanks,
     }
 )(HomePage)
